feat: move base types on a top level
